square

+++ includes/canvas.h
@@ -6,7 +6,7 @@
 /*   By: lmartin <lmartin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/29 07:18:35 by lmartin           #+#    #+#             */
-/*   Updated: 2019/10/29 08:35:05 by lmartin          ###   ########.fr       */
+/*   Updated: 2019/11/06 03:05:06 by lmartin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,6 +22,6 @@ typedef	struct	t_canvas
 	float	distance;
 }				s_canvas;
 
-s_canvas		*new_canvas(int width, int heigth, int distance);
+s_canvas		*new_canvas(float width, float heigth, float distance);
 
 #endif
