+++ srcs/canvas.c
@@ -6,13 +6,13 @@
 /*   By: lmartin <lmartin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/29 07:19:55 by lmartin           #+#    #+#             */
-/*   Updated: 2019/10/29 08:30:10 by lmartin          ###   ########.fr       */
+/*   Updated: 2019/11/06 03:04:54 by lmartin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "canvas.h"
 
-s_canvas		*new_canvas(int width, int heigth, int distance)
+s_canvas		*new_canvas(float width, float heigth, float distance)
 {
 	s_canvas	*canvas;
